requirements.txt
pyyaml
asf-search 
pyproj 
scipy 
rasterio 
geopandas
boto3
dem-stitcher